fix prefix-suffix

+++ jolt-core/src/zkvm/lookup_table/suffixes/sign_extension.rs
@@ -9,6 +9,6 @@ impl<const WORD_SIZE: usize> SparseDenseSuffix for SignExtensionSuffix<WORD_SIZE
         let (_, y) = b.uninterleave();
         let padding_len = std::cmp::min(u64::from(y).trailing_zeros() as usize, y.len());
         // 0b11...100...0
-        (1u64 << WORD_SIZE) - (1u64 << (WORD_SIZE - padding_len))
+        ((1u128 << WORD_SIZE) - (1u128 << (WORD_SIZE - padding_len))) as u64
     }
 }
